Extract formatting logic
